Fix database schema and PO features - change all IDs to BIGINT UNSIGNED, fix PO reference duplication, add date validation, auto-fill unit costs

+++ backend/src/database/sia.sql
@@ -4,7 +4,7 @@ USE cecilles_nstyle_db;
 SET FOREIGN_KEY_CHECKS = 0;
 
 CREATE TABLE IF NOT EXISTS `users` (
-	`id` INT UNSIGNED AUTO_INCREMENT PRIMARY KEY,
+	`id` BIGINT UNSIGNED AUTO_INCREMENT PRIMARY KEY,
 	`username` VARCHAR(100) UNIQUE,
 	`email` VARCHAR(255) UNIQUE,
 	`password_hash` VARCHAR(255) NOT NULL,
@@ -15,44 +15,44 @@ CREATE TABLE IF NOT EXISTS `users` (
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4;
 
 CREATE TABLE IF NOT EXISTS `roles` (
-	`id` INT UNSIGNED AUTO_INCREMENT PRIMARY KEY,
+	`id` BIGINT UNSIGNED AUTO_INCREMENT PRIMARY KEY,
 	`name` VARCHAR(100) UNIQUE,
 	`description` VARCHAR(255),
 	`created_at` TIMESTAMP DEFAULT CURRENT_TIMESTAMP
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4;
 
 CREATE TABLE IF NOT EXISTS `permissions` (
-	`id` INT UNSIGNED AUTO_INCREMENT PRIMARY KEY,
+	`id` BIGINT UNSIGNED AUTO_INCREMENT PRIMARY KEY,
 	`name` VARCHAR(150) UNIQUE,
 	`description` VARCHAR(255)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4;
 
 CREATE TABLE IF NOT EXISTS `role_permissions` (
-	`role_id` INT UNSIGNED NOT NULL,
-	`permission_id` INT UNSIGNED NOT NULL,
+	`role_id` BIGINT UNSIGNED NOT NULL,
+	`permission_id` BIGINT UNSIGNED NOT NULL,
 	PRIMARY KEY (`role_id`,`permission_id`),
 	FOREIGN KEY (`role_id`) REFERENCES `roles`(`id`) ON DELETE CASCADE,
 	FOREIGN KEY (`permission_id`) REFERENCES `permissions`(`id`) ON DELETE CASCADE
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4;
 
 CREATE TABLE IF NOT EXISTS `user_roles` (
-	`user_id` INT UNSIGNED NOT NULL,
-	`role_id` INT UNSIGNED NOT NULL,
+	`user_id` BIGINT UNSIGNED NOT NULL,
+	`role_id` BIGINT UNSIGNED NOT NULL,
 	PRIMARY KEY (`user_id`,`role_id`),
 	FOREIGN KEY (`user_id`) REFERENCES `users`(`id`) ON DELETE CASCADE,
 	FOREIGN KEY (`role_id`) REFERENCES `roles`(`id`) ON DELETE CASCADE
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4;
 
 CREATE TABLE IF NOT EXISTS `user_permissions` (
-	`user_id` INT UNSIGNED NOT NULL,
-	`permission_id` INT UNSIGNED NOT NULL,
+	`user_id` BIGINT UNSIGNED NOT NULL,
+	`permission_id` BIGINT UNSIGNED NOT NULL,
 	PRIMARY KEY (`user_id`,`permission_id`),
 	FOREIGN KEY (`user_id`) REFERENCES `users`(`id`) ON DELETE CASCADE,
 	FOREIGN KEY (`permission_id`) REFERENCES `permissions`(`id`) ON DELETE CASCADE
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4;
 
 CREATE TABLE IF NOT EXISTS `categories` (
-	`id` INT UNSIGNED AUTO_INCREMENT PRIMARY KEY,
+	`id` BIGINT UNSIGNED AUTO_INCREMENT PRIMARY KEY,
 	`name` VARCHAR(150) NOT NULL,
 	`description` TEXT,
 	`created_at` TIMESTAMP DEFAULT CURRENT_TIMESTAMP
@@ -64,7 +64,7 @@ CREATE TABLE IF NOT EXISTS `products` (
 	`name` VARCHAR(255) NOT NULL,
 	`brand` VARCHAR(255),
 	`description` TEXT,
-	`category_id` INT UNSIGNED,
+	`category_id` BIGINT UNSIGNED,
 	`price` DECIMAL(12,2) DEFAULT 0.00,
 	`cost` DECIMAL(12,2) DEFAULT 0.00,
 	`stock_quantity` INT DEFAULT 0,
@@ -86,7 +86,7 @@ CREATE TABLE IF NOT EXISTS `inventory_transactions` (
 	`quantity` INT NOT NULL,
 	`location` VARCHAR(255),
 	`reference` VARCHAR(255),
-	`user_id` INT UNSIGNED,
+	`user_id` BIGINT UNSIGNED,
 	`reason` TEXT,
 	`balance_after` INT,
 	`created_at` TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
@@ -99,7 +99,7 @@ CREATE TABLE IF NOT EXISTS `damaged_inventory` (
 	`product_id` BIGINT UNSIGNED NOT NULL,
 	`quantity` INT NOT NULL,
 	`reason` TEXT,
-	`reported_by` INT UNSIGNED,
+	`reported_by` BIGINT UNSIGNED,
 	`created_at` TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
 	FOREIGN KEY (`product_id`) REFERENCES `products`(`id`) ON DELETE CASCADE,
 	FOREIGN KEY (`reported_by`) REFERENCES `users`(`id`) ON DELETE SET NULL
@@ -150,7 +150,7 @@ CREATE TABLE IF NOT EXISTS `sales` (
 	`id` BIGINT UNSIGNED AUTO_INCREMENT PRIMARY KEY,
 	`sale_number` VARCHAR(100) UNIQUE,
 	`date` TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-	`clerk_id` INT UNSIGNED,
+	`clerk_id` BIGINT UNSIGNED,
 	`customer_id` BIGINT UNSIGNED,
 	`subtotal` DECIMAL(12,2) DEFAULT 0.00,
 	`tax` DECIMAL(12,2) DEFAULT 0.00,
@@ -175,7 +175,7 @@ CREATE TABLE IF NOT EXISTS `sale_items` (
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4;
 
 CREATE TABLE IF NOT EXISTS `employees` (
-	`id` INT UNSIGNED AUTO_INCREMENT PRIMARY KEY,
+	`id` BIGINT UNSIGNED AUTO_INCREMENT PRIMARY KEY,
 	`name` VARCHAR(255) NOT NULL,
 	`role` VARCHAR(100),
 	`contact_type` VARCHAR(50),
@@ -189,7 +189,7 @@ CREATE TABLE IF NOT EXISTS `employees` (
 
 CREATE TABLE IF NOT EXISTS `attendance` (
 	`id` BIGINT UNSIGNED AUTO_INCREMENT PRIMARY KEY,
-	`employee_id` INT UNSIGNED NOT NULL,
+	`employee_id` BIGINT UNSIGNED NOT NULL,
 	`date` DATE NOT NULL,
 	`clock_in` TIME,
 	`clock_out` TIME,
@@ -200,7 +200,7 @@ CREATE TABLE IF NOT EXISTS `attendance` (
 
 CREATE TABLE IF NOT EXISTS `payrolls` (
 	`id` BIGINT UNSIGNED AUTO_INCREMENT PRIMARY KEY,
-	`employee_id` INT UNSIGNED NOT NULL,
+	`employee_id` BIGINT UNSIGNED NOT NULL,
 	`period_start` DATE,
 	`period_end` DATE,
 	`gross_pay` DECIMAL(12,2) DEFAULT 0.00,
@@ -208,7 +208,7 @@ CREATE TABLE IF NOT EXISTS `payrolls` (
 	`advances` DECIMAL(12,2) DEFAULT 0.00,
 	`net_pay` DECIMAL(12,2) DEFAULT 0.00,
 	`status` ENUM('PENDING','PROCESSED','PAID') DEFAULT 'PENDING',
-	`processed_by` INT UNSIGNED,
+	`processed_by` BIGINT UNSIGNED,
 	`processed_at` TIMESTAMP NULL,
 	FOREIGN KEY (`employee_id`) REFERENCES `employees`(`id`) ON DELETE CASCADE,
 	FOREIGN KEY (`processed_by`) REFERENCES `users`(`id`) ON DELETE SET NULL
@@ -218,14 +218,14 @@ CREATE TABLE IF NOT EXISTS `saved_reports` (
 	`id` BIGINT UNSIGNED AUTO_INCREMENT PRIMARY KEY,
 	`name` VARCHAR(255),
 	`filters` JSON,
-	`owner_id` INT UNSIGNED,
+	`owner_id` BIGINT UNSIGNED,
 	`created_at` TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
 	FOREIGN KEY (`owner_id`) REFERENCES `users`(`id`) ON DELETE SET NULL
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4;
 
 CREATE TABLE IF NOT EXISTS `audit_logs` (
 	`id` BIGINT UNSIGNED AUTO_INCREMENT PRIMARY KEY,
-	`user_id` INT UNSIGNED,
+	`user_id` BIGINT UNSIGNED,
 	`action` VARCHAR(255) NOT NULL,
 	`resource_type` VARCHAR(100),
 	`resource_id` VARCHAR(255),
@@ -240,7 +240,7 @@ CREATE TABLE IF NOT EXISTS `files` (
 	`original_name` VARCHAR(255),
 	`type` VARCHAR(50),
 	`size` BIGINT,
-	`uploaded_by` INT UNSIGNED,
+	`uploaded_by` BIGINT UNSIGNED,
 	`uploaded_at` TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
 	FOREIGN KEY (`uploaded_by`) REFERENCES `users`(`id`) ON DELETE SET NULL
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4;
@@ -248,7 +248,7 @@ CREATE TABLE IF NOT EXISTS `files` (
 CREATE TABLE IF NOT EXISTS `notifications` (
 	`id` BIGINT UNSIGNED AUTO_INCREMENT PRIMARY KEY,
 	`type` VARCHAR(100),
-	`recipient_user_id` INT UNSIGNED,
+	`recipient_user_id` BIGINT UNSIGNED,
 	`payload` JSON,
 	`status` ENUM('PENDING','SENT','FAILED') DEFAULT 'PENDING',
 	`sent_at` TIMESTAMP NULL,
